Refactor HTTP config code to support PAC files
